requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
beautifulsoup4>=4.11.0
google-generativeai>=0.3.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
schedule>=1.2.0
aiohttp>=3.8.0
lxml>=4.9.0
python-multipart>=0.0.6